Move resizing to a separate file.
